# Speed up configuration resolution