This speeds up the `getPackageConfigs` call issued by `getDefaultConfiguration` from 14.8 ms to 3.6 ms for the particular project that I'm benchmarking. There is still a lot more to gain, but this is a good start.