@s-ludwig 2024-05-03
Added two more optimizations to get down to 3.1 ms.