@s-ludwig 2024-05-03
Now down to 2.6 ms.