@atilaneves 2024-05-09
> You identified this particular piece of code to be responsible for a considerable slowdown, so I went head and optimized it

Yes, and thanks!

> a 20% total improvement is not what I would associate with "death by a thousand cuts".

Normally, no, but the majority of the runtime is still in `getPackageConfigs`. It gets called twice to make things worse, once to add the test runner configuration and another in the generator when getting the build information (so with different arguments). Our build has 80 dub packages, which means death by 160 cuts. If each "cut" is ~50ms on average it means an added runtime of 8s.