@s-ludwig 2024-05-10
> Normally, no, but the majority of the runtime is still in getPackageConfigs. It gets called twice to make things worse, once to add the test runner configuration and another in the generator when getting the build information (so with different arguments). Our build has 80 dub packages, which means death by 160 cuts. If each "cut" is ~50ms on average it means an added runtime of 8s.

But no matter how often `getPackageConfigs` gets called, a 2x improvement stays a 2x improvement. The death by a thousand cuts would IMO apply more to a situation where a thousand different places each contribute a thousandth of the runtime.

Just to quickly get the numbers straight here:
- The overall number for you improved by roughly 25% - assuming that `getPackageConfigs` takes up 50% of the total time after the optimization,  that would mean a speedup for `getPackageConfigs` of only 1.5x
- Assuming the 160 calls per total runtime is true, that means with the optimized version a single call takes about 38 ms on average (15x more than for my benchmark project)
- Your pathological test case improved by roughly 3.2x
- In my test case, I'm getting an improvement of around 5.7x
- Applying the 3.2x factor above to estimate the overall runtime taken by `getPackageConfigs` would result in just 29% of the total runtime of 15 s (11% afterwards for the 12 s)

What seems odd here is that, speaking of the pathological test case, you both seem to have a larger payload (assuming the single-thread performance of the machine is not much lower) and a considerably worse speedup at the same time. I would usually expect the opposite to be true and it would be interesting to find out why that happens.

On the other hand, if I should bet, I would say that the reality is that the majority of the runtime is actually spent somewhere else. Taking the estimation from the last bullet point, this would mean that optimizing `getPackageConfigs` alone will never result in less than maybe 11 s. Did you measure absolute time, or is this a statistical measurement from a sampling profiler (which can be heavily skewed)?